ScreenGridLayer: devide into Composite and SubLayer
