+++ modules/aggregation-layers/src/screen-grid-layer/screen-grid-layer.js
@@ -18,65 +18,38 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import {Layer, WebMercatorViewport, createIterable, log, experimental} from '@deck.gl/core';
+import {
+  CompositeLayer,
+  WebMercatorViewport,
+  createIterable,
+  log,
+  experimental
+} from '@deck.gl/core';
 const {count} = experimental;
-import {defaultColorRange, colorRangeToFlatArray} from '../utils/color-utils';
 import GPUGridAggregator from '../utils/gpu-grid-aggregation/gpu-grid-aggregator';
 import {AGGREGATION_OPERATION} from '../utils/aggregation-operation-utils';
+import ScreenGridCellLayer from './screen-grid-cell-layer';
 
 import GL from '@luma.gl/constants';
-import {Model, Geometry, Buffer, FEATURES, hasFeatures} from '@luma.gl/core';
+import {Buffer} from '@luma.gl/core';
 
-import vs from './screen-grid-layer-vertex.glsl';
-import fs from './screen-grid-layer-fragment.glsl';
-
-const DEFAULT_MINCOLOR = [0, 0, 0, 0];
-const DEFAULT_MAXCOLOR = [0, 255, 0, 255];
-const COLOR_PROPS = [`minColor`, `maxColor`, `colorRange`, `colorDomain`];
-
-const defaultProps = {
-  cellSizePixels: {value: 100, min: 1},
-  cellMarginPixels: {value: 2, min: 0, max: 5},
-
-  colorDomain: null,
-  colorRange: defaultColorRange,
-
+const defaultProps = Object.assign({}, ScreenGridCellLayer.defaultProps, {
   getPosition: {type: 'accessor', value: d => d.position},
   getWeight: {type: 'accessor', value: d => [1, 0, 0]},
 
   gpuAggregation: true,
   aggregation: 'SUM'
-};
+});
 
-export default class ScreenGridLayer extends Layer {
-  getShaders() {
-    return super.getShaders({vs, fs, modules: ['picking']});
-  }
-
+export default class ScreenGridLayer extends CompositeLayer {
   initializeState() {
-    const attributeManager = this.getAttributeManager();
     const {gl} = this.context;
-    if (!hasFeatures(gl, [FEATURES.TEXTURE_FLOAT])) {
+    if (!ScreenGridCellLayer.isSupported(gl)) {
       // max aggregated value is sampled from a float texture
       this.setState({supported: false});
       log.error(`ScreenGridLayer: ${this.id} is not supported on this browser`)();
       return;
     }
-
-    attributeManager.addInstanced({
-      instancePositions: {size: 3, update: this.calculateInstancePositions},
-      instanceCounts: {
-        size: 4,
-        transition: true,
-        accessor: ['getPosition', 'getWeight'],
-        update: this.calculateInstanceCounts,
-        noAlloc: true
-      }
-    });
-
-    const options = {
-      id: `${this.id}-aggregator`
-    };
     const weights = {
       color: {
         size: 1,
@@ -86,8 +59,7 @@ export default class ScreenGridLayer extends Layer {
     };
     this.setState({
       supported: true,
-      model: this._getModel(gl),
-      gpuGridAggregator: new GPUGridAggregator(gl, options),
+      gpuGridAggregator: new GPUGridAggregator(gl, {id: `${this.id}-aggregator`}),
       weights
     });
   }
@@ -98,8 +70,6 @@ export default class ScreenGridLayer extends Layer {
 
   updateState(opts) {
     super.updateState(opts);
-
-    this._updateUniforms(opts);
 
     if (opts.changeFlags.dataChanged) {
       this._processData();
@@ -118,6 +88,33 @@ export default class ScreenGridLayer extends Layer {
     }
   }
 
+  renderLayers() {
+    if (!this.state.supported) {
+      return [];
+    }
+    const {weights, maxTexture, numRow, numCol, numInstances} = this.state;
+    const {updateTriggers} = this.props;
+
+    return new ScreenGridCellLayer(
+      this.props,
+      this.getSubLayerProps({
+        id: 'cell-layer',
+        updateTriggers
+      }),
+      {
+        data: {
+          attributes: {
+            instanceCounts: weights.color.aggregationBuffer
+          }
+        },
+        maxTexture,
+        numRow,
+        numCol,
+        numInstances
+      }
+    );
+  }
+
   finalizeState() {
     super.finalizeState();
 
@@ -132,63 +129,6 @@ export default class ScreenGridLayer extends Layer {
     if (maxTexture) {
       maxTexture.delete();
     }
-  }
-
-  draw({uniforms}) {
-    if (!this.state.supported) {
-      return;
-    }
-    const {parameters = {}} = this.props;
-    const minColor = this.props.minColor || DEFAULT_MINCOLOR;
-    const maxColor = this.props.maxColor || DEFAULT_MAXCOLOR;
-
-    // If colorDomain not specified we use default domain [1, maxCount]
-    // maxCount value will be sampled form maxTexture in vertex shader.
-    const colorDomain = this.props.colorDomain || [1, 0];
-    const {model, cellScale, shouldUseMinMax, colorRange, maxTexture} = this.state;
-    const layerUniforms = {
-      minColor,
-      maxColor,
-      maxTexture,
-      cellScale,
-      colorRange,
-      colorDomain,
-      shouldUseMinMax
-    };
-
-    uniforms = Object.assign(layerUniforms, uniforms);
-    model.draw({
-      uniforms,
-      parameters: Object.assign(
-        {
-          depthTest: false,
-          depthMask: false
-        },
-        parameters
-      )
-    });
-  }
-
-  calculateInstancePositions(attribute, {numInstances}) {
-    const {width, height} = this.context.viewport;
-    const {cellSizePixels} = this.props;
-    const {numCol} = this.state;
-    const {value, size} = attribute;
-
-    for (let i = 0; i < numInstances; i++) {
-      const x = i % numCol;
-      const y = Math.floor(i / numCol);
-      value[i * size + 0] = ((x * cellSizePixels) / width) * 2 - 1;
-      value[i * size + 1] = 1 - ((y * cellSizePixels) / height) * 2;
-      value[i * size + 2] = 0;
-    }
-  }
-
-  calculateInstanceCounts(attribute, {numInstances}) {
-    const {aggregationBuffer} = this.state;
-    attribute.update({
-      buffer: aggregationBuffer
-    });
   }
 
   getPickingInfo({info, mode}) {
@@ -208,12 +148,11 @@ export default class ScreenGridLayer extends Layer {
     return info;
   }
 
-  // HELPER Methods
+  // Private Methods
 
   _getAggregationChangeFlags({oldProps, props, changeFlags}) {
-    const cellSizeChanged =
-      props.cellSizePixels !== oldProps.cellSizePixels ||
-      props.cellMarginPixels !== oldProps.cellMarginPixels;
+    const cellSizeChanged = props.cellSizePixels !== oldProps.cellSizePixels;
+    // props.cellMarginPixels !== oldProps.cellMarginPixels; // _TODO_ why checking margin pixels
     const dataChanged = changeFlags.dataChanged || props.aggregation !== oldProps.aggregation;
     const viewportChanged = changeFlags.viewportChanged;
 
@@ -222,22 +161,6 @@ export default class ScreenGridLayer extends Layer {
     }
 
     return null;
-  }
-
-  _getModel(gl) {
-    return new Model(
-      gl,
-      Object.assign({}, this.getShaders(), {
-        id: this.props.id,
-        geometry: new Geometry({
-          drawMode: GL.TRIANGLE_FAN,
-          attributes: {
-            positions: new Float32Array([0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0])
-          }
-        }),
-        isInstanced: true
-      })
-    );
   }
 
   // Process 'data' and build positions and weights Arrays.
@@ -271,23 +194,7 @@ export default class ScreenGridLayer extends Layer {
     this.setState({positions, pointCount});
   }
 
-  _shouldUseMinMax() {
-    const {minColor, maxColor, colorDomain, colorRange} = this.props;
-    if (minColor || maxColor) {
-      log.deprecated('ScreenGridLayer props: minColor and maxColor', 'colorRange, colorDomain')();
-      return true;
-    }
-    // minColor and maxColor not supplied, check if colorRange or colorDomain supplied.
-    // NOTE: colorDomain and colorRange are experimental features, use them only when supplied.
-    if (colorDomain || colorRange) {
-      return false;
-    }
-    // None specified, use default minColor and maxColor
-    return true;
-  }
-
   _updateAggregation(changeFlags) {
-    const attributeManager = this.getAttributeManager();
     const {cellSizePixels, gpuAggregation} = this.props;
 
     const {positions, weights} = this.state;
@@ -318,41 +225,10 @@ export default class ScreenGridLayer extends Layer {
       gridTransformMatrix
     });
 
-    attributeManager.invalidate('instanceCounts');
     this.setState({maxTexture: results.color.maxTexture});
   }
 
-  _updateUniforms({oldProps, props, changeFlags}) {
-    const newState = {};
-    if (COLOR_PROPS.some(key => oldProps[key] !== props[key])) {
-      newState.shouldUseMinMax = this._shouldUseMinMax();
-    }
-
-    if (oldProps.colorRange !== props.colorRange) {
-      newState.colorRange = colorRangeToFlatArray(props.colorRange);
-    }
-
-    if (
-      oldProps.cellMarginPixels !== props.cellMarginPixels ||
-      oldProps.cellSizePixels !== props.cellSizePixels ||
-      changeFlags.viewportChanged
-    ) {
-      const {width, height} = this.context.viewport;
-      const {cellSizePixels, cellMarginPixels} = this.props;
-      const margin = cellSizePixels > cellMarginPixels ? cellMarginPixels : 0;
-
-      newState.cellScale = new Float32Array([
-        ((cellSizePixels - margin) / width) * 2,
-        (-(cellSizePixels - margin) / height) * 2,
-        1
-      ]);
-    }
-    this.setState(newState);
-  }
-
   _updateGridParams() {
-    const attributeManager = this.getAttributeManager();
-    attributeManager.invalidateAll();
     const {width, height} = this.context.viewport;
     const {cellSizePixels} = this.props;
     const {gl} = this.context;
